Fixed a bug in a constructor that caused the edit method to act erratically, removed the presets so they go directly into designedlistview, added addbutton

+++ src/viewmodel/ShirtCreatorViewModel.java
@@ -160,12 +160,15 @@ public class ShirtCreatorViewModel {
 	 */
 	public boolean deleteShirt() {
 
-		Shirt shirtToDelete = this.formShirt();
+		if (!this.database.values().isEmpty() && !this.nameProperty.getValue().isBlank()) {
+			Shirt shirtToDelete = this.formShirt();
 
-		if (this.database.removeByKey(shirtToDelete.hashCode())) {
-			this.update();
-			this.clearTextFields();
-			return true;
+			if (this.database.removeByKey(shirtToDelete.hashCode())) {
+
+				this.update();
+				this.clearTextFields();
+				return true;
+			}
 		}
 
 		return false;
@@ -179,41 +182,39 @@ public class ShirtCreatorViewModel {
 	 */
 	public boolean editShirt() {
 
-		Shirt shirtToEdit = this.formShirt();
+		if (!this.database.values().isEmpty() && !this.nameProperty.getValue().isBlank()) {
 
-		if (this.database.containsKey(shirtToEdit.hashCode())) {
+			Shirt shirtToEdit = this.formShirt();
+			int hash = shirtToEdit.hashCode();
 
-			this.database.findByKey(shirtToEdit.hashCode()).setHasPocket(shirtToEdit.hasPocket());
+			if (this.database.containsKey(shirtToEdit.hashCode())) {
 
-			this.database.findByKey(shirtToEdit.hashCode()).setSize(shirtToEdit.getSize());
+				this.database.replaceByKey(hash, shirtToEdit);
 
-			this.database.findByKey(shirtToEdit.hashCode()).setSleeveLength(shirtToEdit.getSleeveLength());
-			this.database.findByKey(shirtToEdit.hashCode()).setColor(shirtToEdit.getColor());
-			this.database.findByKey(shirtToEdit.hashCode()).setNeckStyle(shirtToEdit.getNeckStyle());
-			this.database.findByKey(shirtToEdit.hashCode()).setMaterial(shirtToEdit.getMaterial());
-			this.database.findByKey(shirtToEdit.hashCode()).setShirtText(shirtToEdit.getShirtText());
+				this.update();
 
-			this.update();
-			this.clearTextFields();
-			return true;
+				this.clearTextFields();
+				return true;
+			}
 		}
-
 		return false;
+
 	}
 
 	/**
 	 * Updates the list property to reflect any changes in the database.
 	 */
 	private void update() {
+
 		this.listProperty.set(FXCollections.observableArrayList(this.database.values()));
+
 	}
 
 	private Shirt formShirt() {
 
-		Shirt shirt = new TShirt(
-			    nameProperty.get(), pocketProperty.get(), shoulderProperty.get(), sizeProperty.get(),
-			    sleeveProperty.get(), colorProperty.get(), neckStyleProperty.get(), materialProperty.get(),
-			    backLengthProperty.get(), textProperty.get());
+		Shirt shirt = new TShirt(nameProperty.get(), pocketProperty.get(), shoulderProperty.get(), sizeProperty.get(),
+				sleeveProperty.get(), colorProperty.get(), neckStyleProperty.get(), materialProperty.get(),
+				backLengthProperty.get(), textProperty.get());
 
 		return shirt;
 	}
@@ -226,6 +227,7 @@ public class ShirtCreatorViewModel {
 		nameProperty.set("");
 		shoulderProperty.set(null);
 		sizeProperty.set(null);
+		this.sleeveProperty.set(null);
 		backLengthProperty.set(null);
 		colorProperty.set(null);
 		neckStyleProperty.set(null);
